requirements.txt
python-dotenv
streamlit
openai
chromadb
streamlit_chat